% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfr_command.R
\name{sfr_command}
\alias{sfr_command}
\title{Title}
\usage{
sfr_command(url, output_folder = NULL, timestamped_output = NULL,
  format = "csv", save_crawl_file = TRUE, export_tabs = NULL,
  export_bulk = NULL, export_report = NULL, overwrite = TRUE,
  headless = TRUE, config = NULL, create_sitemap = FALSE,
  create_images_sitemap = FALSE, use_majestic = NULL,
  use_mozscape = NULL, use_ahrefs = NULL,
  use_google_analytics = NULL, use_google_search_console = NULL)
}
\arguments{
\item{use_google_search_console}{}
}
\description{
Title
}
