% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfr_load_all_reports.R
\name{sfr_load_all_reports}
\alias{sfr_load_all_reports}
\title{Load all exported reports, bulk exports or tabs into enviroment as
    properly named data frames}
\usage{
sfr_load_all_reports(output_path)
}
\arguments{
\item{output_path}{character string. The folder path of directory
that contains exported spreadsheets}
}
\value{
data frames from all the spreadsheet files.
}
\description{
Load all exported reports, bulk exports or tabs into enviroment as
    properly named data frames
}
