% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfr_setup_windows.R
\name{sfr_setup_windows}
\alias{sfr_setup_windows}
\title{Function to setup the Screaming Frog on Windows systems}
\usage{
sfr_setup_windows(path)
}
\arguments{
\item{path}{the path to the folder of Screaming Frog. Usually similar to
'C:/Program Files/Screaming Frog SEO Spider/'.}
}
\value{
invisibly adds the folder to enviromental variable
}
\description{
Function to setup the Screaming Frog on Windows systems
}
